+++ scripts/git-subtree/open-api/push.ps1
@@ -1,0 +1,4 @@
+git subtree push `
+    -P workspace/plugins/open-api `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-open-api.git `
+    main